Build remodeling after the system in the Go main source.
Now using the script from Go to build pkg dependencies automatically. Moved
pkg files under a subdirectory as they would be in the main pkg tree, this is
also needed for the dependecy script to work right. Decreased the load on the
root Makefile, it now delegates package tasks to the pkg dir Makefile.

Dependencies between packages and commands are still kludged instead of
handled properly.