Rewrote Bubble Babble decode, less lookahead.
The new decode method doesn't require that the end of the decode data is the
end of the Babble string. It reads the data until it has read a complete
Babble string, then returns that. The error messages should also always note a
precise place where the stream is noticed to no longer match a valid Babble
string.