Manager replaces world as root singleton.
World is now a component handler within manager (it's not actually handling
any specific components, but it still makes sense as a app state component
itself). The world global variable is gone, as are the straggling references
to it around the codebase. Saving games is now done via the manager, using the
manager's toplevel serialization framework.